+++ cargoLoader.Domain/Models/Cargo.cs
@@ -11,16 +11,16 @@ namespace CargoLoader.Domain.Models
     {
         public string Marking { get; set; } = null!;
         public string Name { get; set; } = null!;
-        public double Width { get; set; }
-        public double Height { get; set; }
-        public double Length { get; set; }
-        public double? Volume { get; set; }
-        public double Weight { get; set; }
+        public decimal Width { get; set; }
+        public decimal Height { get; set; }
+        public decimal Length { get; set; }
+        public decimal? Volume { get; set; }
+        public decimal Weight { get; set; }
         public bool IsFragile { get; set; }
         public bool IsRotatable { get; set; }
         public bool IsProp { get; set; }
         public bool IsContainer { get; set; }
-        public double? Capacity { get; set; }
+        public decimal? Capacity { get; set; }
         public int OrderId { get; set; }
         public Order Order { get; set; } = null!;
         public List<Cargo>? ContainedCargo { get; set; }
