+++ cargoLoader.Domain/Models/Product.cs
@@ -10,11 +10,11 @@ namespace CargoLoader.Domain.Models
     {
         public string Marking { get; set; }
         public string Name { get; set; }
-        public double Width { get; set; }
-        public double Height { get; set; }
-        public double Length { get; set; }
-        public double? Volume { get; set; }
-        public double Weight { get; set; }
+        public decimal Width { get; set; }
+        public decimal Height { get; set; }
+        public decimal Length { get; set; }
+        public decimal? Volume { get; set; }
+        public decimal Weight { get; set; }
         public bool IsFragile { get; set; }
         public bool IsRotatable { get; set; }
         public bool IsProp { get; set; }
